+++ packages/checkout/widgets/src/definitions/events/swapEvents.ts
@@ -1,20 +1,10 @@
 /**
- * Represents an event object emitted by the Swap Widget.
- * @property {SwapEventType} type - The type of the event.
- * @property {T} data - The data contained in the event.
- */
-export type SwapEvent<T> = {
-  type: SwapEventType;
-  data: T;
-};
-
-/**
  * Enum representing possible Swap Widget event types.
  */
 export enum SwapEventType {
+  CLOSE_WIDGET = 'close-widget',
   SUCCESS = 'success',
   FAILURE = 'failure',
-  CLOSE_WIDGET = 'close-widget',
   REJECTED = 'rejected',
 }